requirements.txt
PdfReader
python-dotenv
llama-index
llama-index-llms-azure-openai
llama-index-embeddings-azure-openai
docx2txt